[OS_ARCH] rename directoy and maven modules
* properly filter for architecture
