+++ app/pom.xml
@@ -28,19 +28,19 @@
 
     <dependency>
       <groupId>de.bmarwell.jtop</groupId>
-      <artifactId>jtop-library-aix</artifactId>
+      <artifactId>jtop-library-aix-ppc64</artifactId>
       <version>1.0-SNAPSHOT</version>
       <scope>runtime</scope>
     </dependency>
     <dependency>
       <groupId>de.bmarwell.jtop</groupId>
-      <artifactId>jtop-library-linux</artifactId>
+      <artifactId>jtop-library-linux-noarch</artifactId>
       <version>1.0-SNAPSHOT</version>
       <scope>runtime</scope>
     </dependency>
     <dependency>
       <groupId>de.bmarwell.jtop</groupId>
-      <artifactId>jtop-library-osx</artifactId>
+      <artifactId>jtop-library-darwin-aarch64</artifactId>
       <version>1.0-SNAPSHOT</version>
       <scope>runtime</scope>
     </dependency>
