+++ library/aix-ppc64/src/main/java/de/bmarwell/jtop/lib/aix/AixProcess.java
