+++ library/darwin-aarch64/src/main/java/de/bmarwell/jtop/lib/osx/ffm/loadavg.java
