+++ library/darwin-aarch64/src/main/java/de/bmarwell/jtop/lib/osx/ffm/proc_bsdshortinfo.java
