+++ library/linux-noarch/src/main/java/de/bmarwell/jtop/lib/linux/LinuxProcess.java
