--- library/osx/pom.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>de.bmarwell.jtop</groupId>
-        <artifactId>jtop-library</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-
-    <groupId>de.bmarwell.jtop</groupId>
-    <artifactId>jtop-library-osx</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <name>jtop :: library :: osx</name>
-
-    <dependencies>
-        <dependency>
-            <groupId>de.bmarwell.jtop</groupId>
-            <artifactId>jtop-library-api</artifactId>
-            <version>1.0-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
-    </dependencies>
-
-
-</project>
